# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: May Chicken & Burger Restaurant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/categories:
    get:
      operationId: listAdminCategories
      tags: [admin]
      summary: List all categories (admin)
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [admin]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /admin/categories/{id}:
    patch:
      operationId: updateCategory
      tags: [admin]
      summary: Update a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [admin]
      summary: Delete a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── ADMIN – PRODUCTS ──────────────────────────────────────────────────────
  /admin/items:
    get:
      operationId: listAdminItems
      tags: [admin]
      summary: List all menu items (admin)
      parameters:
        - name: categoryId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Menu items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"
    post:
      operationId: createAdminMenuItem
      tags: [admin]
      summary: Create a menu item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
  /admin/items/{id}:
    patch:
      operationId: updateAdminMenuItem
      tags: [admin]
      summary: Update a menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuItemUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
    delete:
      operationId: deleteAdminMenuItem
      tags: [admin]
      summary: Delete a menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── ADMIN – ORDERS ────────────────────────────────────────────────────────
  /admin/orders:
    get:
      operationId: listAdminOrders
      tags: [admin]
      summary: List all orders (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, preparing, ready, delivering, completed, cancelled]
        - name: orderType
          in: query
          schema:
            type: string
            enum: [delivery, pickup]
        - name: date
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
  /admin/orders/{id}:
    get:
      operationId: getAdminOrder
      tags: [admin]
      summary: Get a single order (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    patch:
      operationId: updateAdminOrder
      tags: [admin]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusPatch"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # ── ADMIN – CUSTOMERS ─────────────────────────────────────────────────────
...
  /admin/opening-hours:
    get:
      operationId: getAdminOpeningHours
      tags: [admin]
      summary: Get opening hours (admin)
      responses:
        "200":
          description: Opening hours
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DayHours"
    put:
      operationId: updateOpeningHours
      tags: [admin]
      summary: Update opening hours for all days
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpeningHoursUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DayHours"

  # ── ADMIN – COUPONS ───────────────────────────────────────────────────────
  /admin/coupons:
    get:
      operationId: listCoupons
      tags: [admin]
      summary: List all coupons
      responses:
        "200":
          description: Coupons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Coupon"
    post:
      operationId: createCoupon
      tags: [admin]
      summary: Create a coupon
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
  /admin/coupons/{id}:
    patch:
      operationId: updateCoupon
      tags: [admin]
      summary: Update a coupon
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
    delete:
      operationId: deleteCoupon
      tags: [admin]
      summary: Delete a coupon
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── ADMIN – SETTINGS ─────────────────────────────────────────────────────
  /admin/settings:
    get:
      operationId: getAdminSettings
      tags: [admin]
      summary: Get all app settings
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppSettings"
    put:
      operationId: updateAdminSettings
      tags: [admin]
      summary: Update app settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettingsUpdate"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppSettings"

  # ── KITCHEN ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Category:
      type: object
      required: [id, name, slug, sortOrder, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        sortOrder:
          type: integer
        itemCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    CategoryInput:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
          minLength: 1
        slug:
          type: string
          minLength: 1
        description:
          type: string
        imageUrl:
          type: string
        sortOrder:
          type: integer
    CategoryUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        slug:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        sortOrder:
          type: integer

    # ── MENU ITEM ─────────────────────────────────────────────────────────
    MenuItem:
      type: object
      required: [id, name, price, categoryId, available, featured, sortOrder, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        categoryId:
          type: integer
        available:
          type: boolean
        featured:
          type: boolean
        imageUrl:
          type: ["string", "null"]
        sortOrder:
          type: integer
        category:
          $ref: "#/components/schemas/Category"
        createdAt:
          type: string
          format: date-time
    MenuItemInput:
      type: object
      required: [name, price, categoryId]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        price:
          type: number
          minimum: 0
        categoryId:
          type: integer
        available:
          type: boolean
        featured:
          type: boolean
        imageUrl:
          type: string
        sortOrder:
          type: integer
    MenuItemUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        price:
          type: number
          minimum: 0
        categoryId:
          type: integer
        available:
          type: boolean
        featured:
          type: boolean
        imageUrl:
          type: string
        sortOrder:
          type: integer

    # ── RESTAURANT INFO ───────────────────────────────────────────────────
...
    DayHours:
      type: object
      required: [id, dayOfWeek, dayName, isClosed]
      properties:
        id:
          type: integer
        dayOfWeek:
          type: integer
        dayName:
          type: string
        openTime:
          type: ["string", "null"]
        closeTime:
          type: ["string", "null"]
        isClosed:
          type: boolean
...
    OpeningHoursUpdate:
      type: object
      required: [hours]
      properties:
        hours:
          type: array
          items:
            $ref: "#/components/schemas/DayHoursInput"

    # ── DELIVERY AREAS ────────────────────────────────────────────────────
...
    CouponInput:
      type: object
      required: [code, discountType, discountValue]
      properties:
        code:
          type: string
          minLength: 1
        description:
          type: string
        discountType:
          type: string
          enum: [percentage, fixed]
        discountValue:
          type: number
          minimum: 0
        minOrder:
          type: number
          minimum: 0
        active:
          type: boolean
        expiresAt:
          type: string
          format: date-time
        maxUsage:
          type: integer
...
    Order:
      type: object
      required: [id, orderNumber, orderType, status, customerName, customerPhone, subtotal, deliveryFee, discountAmount, total, items, createdAt]
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        orderType:
          type: string
          enum: [delivery, pickup]
        status:
          type: string
          enum: [pending, confirmed, preparing, ready, delivering, completed, cancelled]
        customerName:
          type: string
        customerPhone:
          type: string
        customerEmail:
          type: ["string", "null"]
        deliveryAddress:
          type: ["string", "null"]
        postalCode:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        subtotal:
          type: number
        deliveryFee:
          type: number
        discountAmount:
          type: number
        total:
          type: number
        couponCode:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
          format: date-time
    OrderStatusPatch:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, confirmed, preparing, ready, delivering, completed, cancelled]
        notes:
          type: string
...